include/uk: Move compiler definitions to own header
Move compiler definitions from `include/uk/essentials.h`
to `include/uk/compiler.h`, for modularity enhancement
and easier maintenance.
